pyproject.toml
[project]
name = "kvm-mgr-service"
version = "1.0.0"
description = "KVM Manager Service - Python backend for controlling MLEEDA KVM1001A hardware switch"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pyserial>=3.5",
    "psutil>=5.9.6",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "rich>=14.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]
